Leaderboard updates, loader moved to atoms
