+++ src/components/atoms/Loader/loaderTypes.ts
