+++ src/components/atoms/Loader/styles/index.scss
