+++ src/components/molecules/index.ts
@@ -4,7 +4,6 @@ export * from './EmptyCard';
 export * from './EncouragementButton';
 export * from './Input';
 export * from './InstructionCard';
-export * from './Loader';
 export * from './Submission';
 export * from './SubmissionCard';
 export * from './Table';
